requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.8
django-admin-numeric-filter==0.1.6
django-modeltranslation==0.17.3
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.2
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7